Taking Aim tweaks & misc
Changed the "Taking Aim" countdown setting to allow the duration given by the game. A custom duration is still possible but by default it will grab it from the game. If you have a duration that is different from the default duration it will change the setting to "Custom" upon startup.