Fixed Zhaj'hassa's cleansing platforms error report when setting it up
In 40048ba there was introduced OnBossesChanged() call aimed to switch panel visibility without reloading UI.
But there was some big refactoring in dc16258, so OnBossesChanged() got separated for several trial-related files.
But setter for cleansing platform setting wasn't touched, so now it calls for a UI bug (unexisting function).
Still it works fine, but UI error was confusing, and UI reload was required.